fix: move team member helper out of server action
